Moved step 3 to a component
