+++ src/pages/App.tsx
@@ -1,20 +1,21 @@
 /* eslint-disable react/jsx-no-bind */
 /* eslint-disable react/no-array-index-key */
-import React from 'react';
+import React, { useState } from 'react';
 import {
-  Box, Button, Card,
-  CardBody, Grid, Grommet, Header, Heading, Main, Paragraph, Select, TableBody, TableCell,
-  TableRow, Text
+  Box, Button,
+  Grid, Grommet, Header, Heading, Main,
+  Select, TableBody, TableCell,
+  TableRow, Text,
 } from 'grommet';
-import { useEffect, useState } from 'react';
 import { useRecoilState } from 'recoil';
 
 import DataInput from '../components/DataInput';
 import MapDetails from '../components/MapDetails';
 import RenderMap from '../components/RenderMap';
 import StepLabel from '../components/StepLabel';
-import { CHOROPLETH_COLORS, REGIONS, STEPS } from '../constants';
+import { CHOROPLETH_COLORS, STEPS } from '../constants';
 import { recoilState } from '../store';
+import MapRefine from '../components/MapRefine';
 
 const CustomTheme = {
   global: {
@@ -25,7 +26,6 @@ const CustomTheme = {
 };
 
 function App() {
-  const [mismatchedRegionsCount, setMismatchedRegionsCount] = useState(0);
   const [choroplethColorScheme, setChoroplethColorScheme] = useState('Reds');
   const [symbolColorScheme, setSymbolColorScheme] = useState('Red');
   const [symbolShape, setSymbolShape] = useState('Circle');
@@ -34,20 +34,6 @@ function App() {
 
   const symbolColorOptions: string[] = ['Red', 'Blue', 'Green', 'Orange'];
   const symbolShapeOptions: string[] = ['Circle', 'Square', 'Triangle', 'Diamond'];
-
-  useEffect(() => {
-    if (!appState.dataKeys.name) return;
-    fetch(REGIONS[appState.map.region])
-      .then((response) => response.json())
-      .then((geojson) => {
-        const userDataRegions: string[] = appState.userData.data.map((row) => row[appState.dataKeys.name!]);
-        const regions: string[] = geojson.features.map((x: any) => x.properties.name);
-        const mismatchedRegions = userDataRegions.filter(
-          (region) => !regions.includes(region),
-        ) as [];
-        setMismatchedRegionsCount(mismatchedRegions.length);
-      });
-  }, [appState.dataKeys.name]);
 
   function advanceToNext() {
     const currentStep = appState.currentStep + 1;
@@ -65,165 +51,15 @@ function App() {
       userData: { data: [], ready: false },
       currentStep: 0,
       dataKeys: {},
+      mismatchedRegions: 0,
+      choroplethColorScheme: 'Reds',
+      symbolColorScheme: 'Red',
+      symbolShape: 'Circle',
     });
 
     setChoroplethColorScheme('Reds');
     setSymbolColorScheme('Red');
     setSymbolShape('Circle');
-  }
-
-  function renderSymbolMapRefineInputForm() {
-    const columns = Object.keys(appState.userData.data[0]).map((header) => header);
-    return (
-      <Box height="large">
-        <Box pad="medium">
-          <Heading level="4">Time to refine your data</Heading>
-          <TableBody>
-            <TableRow>
-              <TableCell scope="row">
-                <Text alignSelf="start" margin="small">Select column for latitude:</Text>
-              </TableCell>
-              <TableCell>
-                <Select
-                  options={columns}
-                  alignSelf="end"
-                  value={appState.dataKeys.latitude}
-                  onChange={({ option }) => {
-                    setAppState({
-                      ...appState,
-                      dataKeys: { ...appState.dataKeys, latitude: option },
-                    });
-                  }}
-                />
-              </TableCell>
-            </TableRow>
-            <TableRow>
-              <TableCell>
-                <Text alignSelf="start" margin="small">Select column for longitude:</Text>
-              </TableCell>
-              <TableCell>
-                <Select
-                  options={columns}
-                  alignSelf="end"
-                  value={appState.dataKeys.longitude}
-                  onChange={({ option }) => {
-                    setAppState({
-                      ...appState,
-                      dataKeys: { ...appState.dataKeys, longitude: option },
-                    });
-                  }}
-                />
-              </TableCell>
-            </TableRow>
-            <TableRow>
-              <TableCell>
-                <Text alignSelf="start" margin="small">Select column for size:</Text>
-              </TableCell>
-              <TableCell>
-                <Select
-                  options={columns}
-                  alignSelf="end"
-                  value={appState.dataKeys.sizeValues}
-                  onChange={
-                    ({ option }) => setAppState({ ...appState, dataKeys: { ...appState.dataKeys, sizeValues: option }})
-                  }
-                />
-              </TableCell>
-            </TableRow>
-            <TableRow>
-              <TableCell>
-                <Text alignSelf="start" margin="small">Select column for color (optional):</Text>
-              </TableCell>
-              <TableCell>
-                <Select
-                  options={columns}
-                  alignSelf="end"
-                  value={appState.dataKeys.colorValues}
-                  onChange={
-                    ({ option }) => setAppState({ ...appState, dataKeys: { ...appState.dataKeys, colorValues: option }})
-                  }
-                />
-              </TableCell>
-            </TableRow>
-          </TableBody>
-        </Box>
-      </Box>
-    );
-  }
-
-  function renderChoroplethMapRefineInputForm() {
-    const columns = Object.keys(appState.userData.data[0]).map((header) => header);
-    return (
-      <Box height="large">
-        <Box pad="medium">
-          <Heading level="4">Time to refine your data</Heading>
-          <TableBody>
-            <TableRow>
-              <TableCell>
-                <Text alignSelf="start" margin="small">Select column for name:</Text>
-              </TableCell>
-              <TableCell>
-                <Select
-                  options={columns}
-                  value={appState.dataKeys.name}
-                  alignSelf="end"
-                  onChange={({ option }) => setAppState({ ...appState, dataKeys: { ...appState.dataKeys, name: option }})}
-                />
-              </TableCell>
-            </TableRow>
-            <TableRow>
-              <TableCell>
-                <Text alignSelf="start" margin="small">Select column for value:</Text>
-              </TableCell>
-              <TableCell>
-                <Select
-                  options={columns}
-                  value={appState.dataKeys.values}
-                  alignSelf="end"
-                  onChange={({ option }) => setAppState({ ...appState, dataKeys: { ...appState.dataKeys, values: option }})}
-                />
-              </TableCell>
-            </TableRow>
-          </TableBody>
-
-          {appState.dataKeys.name
-          && (
-            <Card margin="small" pad="small" height="small" width="large" background="white">
-              <CardBody pad="xsmall" width="large">
-                { (mismatchedRegionsCount > 0)
-                  ? (
-                    <Paragraph margin="none">
-                      We couldn&#39;t match&nbsp;
-                      {mismatchedRegionsCount}
-                      &nbsp;entries from your file,
-                      &nbsp;your visualization might not be complete.
-                      To resolve this issue, please make sure your data
-                      &nbsp;matches the country names for&nbsp;
-                      {appState.map.region}
-                      .&nbsp;
-                    </Paragraph>
-                  )
-                  : (
-                    <Paragraph margin="none">
-                      Data looks good! All the entries in your map correspond to the set&nbsp;
-                      of country names for&nbsp;
-                      {appState.map.region}
-                      .&nbsp;
-                    </Paragraph>
-                  )}
-                {}
-              </CardBody>
-            </Card>
-          )}
-        </Box>
-      </Box>
-    );
-  }
-
-  function renderRefineStep() {
-    if (appState.map.type === 'Symbol') return renderSymbolMapRefineInputForm();
-
-    return renderChoroplethMapRefineInputForm();
   }
 
   function renderSymbolVisualizeStep() {
@@ -338,7 +174,7 @@ function App() {
               <Box pad="medium">
                 { appState.currentStep === 0 && (<MapDetails />) }
                 { appState.currentStep === 1 && (<DataInput />) }
-                { appState.currentStep === 2 && renderRefineStep() }
+                { appState.currentStep === 2 && (<MapRefine />) }
                 { appState.currentStep === 3 && renderVisualizeStep() }
 
                 <Box direction="row" justify="between">
